improving the cli
